add update version

+++ Common/expr/SimpleExpr.h
@@ -15,11 +15,11 @@
 class SimpleExpr
 {
 public:
-	int CalcuExpr(const char* expr);
+	float CalcuExpr(const char* expr);
 	char* GetNum(char* expr);
 private:
 	std::stack<char> m_stackOpr;
-	std::stack<int> m_stackNum;
+	std::stack<float> m_stackNum;
 	char m_symbol[30];
 };
